requirements/base.txt
flask>=2.0.3
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1

uvicorn==0.20.0
fastapi==0.92.0
jwt==1.3.1
fastapi_auth0==0.3.2
starlette==0.25.0
httpx==0.23.3
